Rename `errors.rs` to `error.rs`
